Improved VScode

+++ sw/simulation/agent.cpp
@@ -1,6 +1,6 @@
 #include "agent.h"
 
-Agent::Agent(int i, const vector<float> &s):
+Agent::Agent(uint8_t i, const vector<float> &s):
 	situation(8,0)
 {
 	this->state = s;
@@ -15,10 +15,9 @@ std::vector<float> Agent::get_states()
 	return state;
 };
 
-float Agent::get_position(int dim)
+float Agent::get_position(uint8_t dim)
 {
-	if (dim < 3)
-	{
+	if (dim < 3) {
 		return state[dim];
 	}
 
@@ -27,7 +26,7 @@ float Agent::get_position(int dim)
 
 void Agent::select_action(){};
 
-int Agent::get_ID()
+uint8_t Agent::get_ID()
 {
 	return ID;
 };
