+++ sw/simulation/agent.h
@@ -4,6 +4,8 @@
 #include <vector>
 #include <stdio.h>
 #include <iostream>
+#include <stdint.h>
+
 #include "controller.h"
 #include "terminalinfo.h"
 
@@ -12,11 +14,12 @@ using namespace std;
 class Agent {
 	vector<float> inputs;
 	vector<float> actions;
+	
 public:
-	Agent(int i, const vector<float> &s); // Make two versions for random initialization
+	Agent(uint8_t i, const vector<float> &s); // Make two versions for random initialization
 	~Agent();
 
-	int ID;
+	uint8_t ID;
 	Controller controller;
 	vector<float> outputs;
 	vector<float> state;
@@ -25,8 +28,8 @@ public:
 	vector<float> get_states();
 	void select_action();
 	virtual void update_position()=0;
-	float get_position(int dim);
-	int get_ID();
+	float get_position(uint8_t dim);
+	uint8_t get_ID();
 };
 
 #endif /*AGENT_H*/
